Openai SDK/basic_agent/pyproject.toml
[project]
name = "basic-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agents>=1.4.0",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
]
